Merge Danbranch with Genebranch (#5)
* Fix Store errors



Remove the main() method from store.py and add missing parentheses on line 11.

* Remove password attribute and method form User

Remove the password from the User class since it was agreed that a password was not needed to shop. 

Removes all asociated methods

* Danbranch (#2)

* Fix Store errors



Remove the main() method from store.py and add missing parentheses on line 11.

* Remove password attribute and method form User

Remove the password from the User class since it was agreed that a password was not needed to shop. 

Removes all asociated methods

* Update Cart and Store class

Update cart class by creating getter and setter methods.
Additional methods added to cart class in order to interact and calculate
within the object.

Remove print statement used for testing in store.py. Remove extraneous newlines in file.

* Add index.py to repo & Changes to user and cart

Add index.py to the project which will be used to run the program.
Necessary changes were made to User and Cart class in order to work with client code (index.py).