Update User, Cart, Store classes and client code
Updates the User class to include rstrip() of user details in setters. Also
changes the history file constant name and references used throughout.

Updates the Cart class to include a new property called taxAmount where a
calculated tax amount is set from within the obejct. Updates the checkout()
method to be able to write to the cart.csv file on a per product in cart basis.
After writing to cart.csv, method writes to transaction.csv for final logging of
tax, subtotal and total nubers.

Updates Store with a new property called name used to store the name of a Store.

Misc. formatting updates throughout all three classes and the client code in order to
standardize the output to the screen